# packages/contracts/api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: AEP BFF API
  version: 1.0.0
servers:
- url: /api/v1
paths:
...
  /projects/{projectName}/files/{path}:
    get:
      operationId: read-file
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: 'File path under specs/ (trailing wildcard: may span multiple segments, i.e. contain
          slashes)'
        in: path
        name: path
        required: true
        schema:
          description: File path under specs/
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileContent'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
      security:
      - userJWT: []
      summary: Read a spec file at HEAD
      tags:
      - Files
  /projects/{projectName}/issues:
    get:
      operationId: list-issues
      summary: List/search GitHub issues on a project's repo
      tags:
      - Issues
      security:
      - userJWT: []
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          type: string
      - description: Comma-separated GitHub labels to filter by
        in: query
        name: labels
        schema:
          type: string
      - description: Keyword search over issue title/body, ranked by distinct-term
          overlap (title weighted double), capped at 25. Recall-biased — used to
          surface related issues before filing a new one.
        in: query
        name: q
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/IssueInfo'
                type: array
                nullable: true
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
    post:
      operationId: create-issue
      summary: Create a GitHub issue on a project's repo
      tags:
      - Issues
      security:
      - userJWT: []
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResult'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
...
  /projects/{projectName}/tasks:
    get:
      operationId: list-tasks
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: Which Tasks to return (default open)
        explode: false
        in: query
        name: state
        schema:
          description: Which Tasks to return (default open)
          enum:
          - open
          - closed
          - all
          type: string
      - description: Filter to Tasks planned from this spec/build version tag (e.g. v3); empty returns
          all versions.
        explode: false
        in: query
        name: tag
        schema:
          description: Filter to Tasks planned from this spec/build version tag (e.g. v3); empty returns
            all versions.
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TaskView'
                type: array
                nullable: true
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
      security:
      - userJWT: []
      summary: List a project's Tasks (live GitHub ⋈ executions → derived status), optionally scoped to
        one spec/build version tag
      tags:
      - Tasks
  /projects/{projectName}/tasks/{issueNumber}:
    get:
      operationId: get-task
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: GitHub issue number of the Task
        in: path
        name: issueNumber
        required: true
        schema:
          description: GitHub issue number of the Task
          format: int64
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
      security:
      - userJWT: []
      summary: Get one Task with its full Execution history
      tags:
      - Tasks
...
  /projects/{projectName}/tasks/{issueNumber}/promote-from-issue:
    post:
      operationId: promote-task-from-issue
      summary: Turn an ad-hoc GitHub issue into a coding Task and dispatch it (async)
      tags:
      - Tasks
      security:
      - userJWT: []
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          type: string
      - description: GitHub issue number to promote and dispatch
        in: path
        name: issueNumber
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoteFromIssueRequest'
        required: true
      responses:
        '202':
          description: Accepted — the funnel dispatches out-of-band.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
...
components:
  schemas:
...
    Error:
      description: Flat error envelope returned by every non-2xx response.
      type: object
      additionalProperties: false
      properties:
        code:
          description: Machine-readable error code (stable slug).
          type: string
          example: validation_failed
        message:
          description: Human-readable error message.
          type: string
          example: request body is invalid
        details:
          description: Field-level details for validation errors.
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
      required:
      - code
      - message
    ErrorDetail:
      description: One field-level validation error.
      type: object
      additionalProperties: false
      properties:
        field:
          description: Location of the offending field (e.g. body.name).
          type: string
        message:
          description: What is wrong with the field.
          type: string
      required:
      - field
      - message
    ExecutionView:
      additionalProperties: false
      properties:
        createdAt:
          format: date-time
          type: string
        endedAt:
          format: date-time
          type: string
          nullable: true
          x-go-type: '*time.Time'
        id:
          type: string
        kind:
          type: string
        reason:
          type: string
        runName:
          type: string
        startedAt:
          format: date-time
          type: string
          nullable: true
          x-go-type: '*time.Time'
        status:
          type: string
      required:
      - id
      - kind
      - status
      - createdAt
      type: object
...
    FileContent:
      additionalProperties: false
      properties:
        content:
          type: string
        path:
          type: string
        sha:
          type: string
      required:
      - path
      - content
      - sha
      type: object
...
    CreateIssueRequest:
      additionalProperties: false
      description: Issue to file on the project's repo. dedupeKey makes creation
        idempotent per open issue (label-encoded), for concurrent alert handlers.
      properties:
        title:
          type: string
        body:
          type: string
        labels:
          items:
            type: string
          type: array
          nullable: true
        dedupeKey:
          type: string
      required:
      - title
      - body
      type: object
    IssueResult:
      additionalProperties: false
      description: Issue metadata after create. deduped=true means an open issue
        with the same dedupeKey already existed — number/url refer to it.
      properties:
        number:
          format: int64
          type: integer
        url:
          type: string
        nodeId:
          type: string
        deduped:
          type: boolean
      required:
      - number
      - url
      - nodeId
      type: object
    IssueInfo:
      additionalProperties: false
      description: 'One issue from list/search. Field names are CAPITALIZED on
        the wire (historical shape the deployed aep-mcp-server parses — do not
        "fix" without a coordinated MCP-server release).'
      properties:
        Number:
          format: int64
          type: integer
        Title:
          type: string
        Body:
          type: string
        URL:
          type: string
        State:
          type: string
        Labels:
          items:
            type: string
          type: array
          nullable: true
      required:
      - Number
      - Title
      - Body
      - URL
      - State
      - Labels
      type: object
...
    Lineage:
      additionalProperties: false
      properties:
        designTag:
          type: string
        specTag:
          type: string
      type: object
...
    PromoteFromIssueRequest:
      additionalProperties: false
      properties:
        componentName:
          description: Component this issue is about
          type: string
      required:
      - componentName
      type: object
...
    TaskDetail:
      additionalProperties: false
      properties:
        blockedBy:
          description: Names of the dependencies this task is waiting on; present when derivedStatus is
            on_hold.
          items:
            type: string
          type: array
          nullable: true
        attention:
          items:
            type: string
          type: array
          nullable: true
        body:
          type: string
        component:
          type: string
        dependsOn:
          items:
            type: string
          type: array
          nullable: true
        derivedStatus:
          type: string
        executionHistory:
          items:
            $ref: '#/components/schemas/ExecutionView'
          type: array
          nullable: true
        executions:
          additionalProperties:
            $ref: '#/components/schemas/ExecutionView'
          type: object
        executorClass:
          type: string
        hold:
          type: boolean
        issueNumber:
          format: int64
          type: integer
        issueUrl:
          type: string
        lineage:
          $ref: '#/components/schemas/Lineage'
        operation:
          type: string
        origin:
          type: string
        rationale:
          type: string
        title:
          type: string
      required:
      - executionHistory
      - issueNumber
      - title
      - issueUrl
      - dependsOn
      - lineage
      - derivedStatus
      - hold
      - attention
      - executions
      type: object
...
    TaskView:
      additionalProperties: false
      properties:
        attention:
          items:
            type: string
          type: array
          nullable: true
        blockedBy:
          description: Names of the dependencies this task is waiting on; present when derivedStatus is
            on_hold.
          items:
            type: string
          type: array
        body:
          type: string
        component:
          type: string
        dependsOn:
          items:
            type: string
          type: array
          nullable: true
        derivedStatus:
          type: string
        executions:
          additionalProperties:
            $ref: '#/components/schemas/ExecutionView'
          type: object
        executorClass:
          type: string
        hold:
          type: boolean
        issueNumber:
          format: int64
          type: integer
        issueUrl:
          type: string
        lineage:
          $ref: '#/components/schemas/Lineage'
        operation:
          type: string
        origin:
          type: string
        rationale:
          type: string
        title:
          type: string
      required:
      - issueNumber
      - title
      - issueUrl
      - dependsOn
      - lineage
      - derivedStatus
      - hold
      - attention
      - executions
      type: object
...
  securitySchemes:
    userJWT:
      bearerFormat: JWT
      description: Thunder end-user JWT (RS256, JWKS-verified)
      scheme: bearer
      type: http